+++ .gitignore
@@ -1,2 +1,3 @@
 !.gitignore
 .venv/
+__pycache__/
