--- Rover Simulation/main.py
@@ -1,1 +1,0 @@
-# For running the simulation main loop
