--- Simulations/Rover-Autonomous-Pathing/Sensors/LimitSwitch.py
@@ -1,3 +1,0 @@
-class LimitSwitch:
-   def detect_edge(self, solar_panel_area, rover_position):
-        return
