--- Simulations/Rover-Autonomous-Pathing/Sensors/RotaryEncoder.py
@@ -1,3 +1,0 @@
-class RotaryEncoder:
-    def getTrackVelocity(self):
-        return
